requirements.txt
streamlit>=1.28.0
feedparser>=6.0.10
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0

